removed isTestMode and inbetween client
